refactor: IntroArrow, IntroText, IntroContainer 스타일 분리
