+++ src/components/intro-text/IntroContainer.tsx
@@ -1,20 +1,7 @@
 /** @jsxImportSource @emotion/react */
-import { css } from '@emotion/react';
+import { introContainerStyle } from './Style';
 import IntroText from './IntroText';
 import IntroArrow from './IntroArrow';
-
-const introContainerStyle = css({
-    color: 'white',
-    display: 'flex',
-    position: 'relative',
-    flexDirection: 'column',
-    '-webkit-box-pack': 'center',
-    justifyContent: 'center',
-    '-webkit-box-align': 'center',
-    alignItems: 'center',
-    height: '100vh',
-    width: 'fit-in-content',
-});
 
 const IntroContainer = () => {
     return (
